requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
anyio==3.7.1
appnope==0.1.3
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
autogen==1.0.16
autopep8==2.0.4
beautifulsoup4==4.12.2
blinker==1.6.3
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi @ file:///private/tmp/cffi-20231014-5216-8jirg3/cffi-1.16.0
charset-normalizer==3.3.1
cleo==2.1.0
click==8.1.7
contourpy==1.2.0
crashtest==0.4.1
cycler==0.12.1
decorator==5.1.1
diskcache==5.6.3
distlib @ file:///private/tmp/python-distlib-20231212-4938-trhlvy/distlib-0.3.8
distro==1.8.0
docopt==0.6.2
docutils==0.20.1
duckdb==0.8.1
dulwich==0.21.7
executing==2.0.1
fastjsonschema==2.19.1
filelock @ file:///private/tmp/python-filelock-20231112-5622-k62fq8/filelock-3.13.1/dist/filelock-3.13.1-py3-none-any.whl#sha256=57dbda9b35157b05fb3e58ee91448612eb674172fab98ee235ccb0b5bee19a1c
FLAML==2.1.1
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.44.0
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
ipython==8.17.2
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
keyring==24.2.0
kiwisolver==1.4.5
lxml==5.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
nh3==0.2.14
numpy==1.26.1
openai==1.3.5
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
pandasai==1.4.5
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.39
protobuf==4.24.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyautogen==0.1.14
pycodestyle==2.11.1
pycparser @ file:///private/tmp/pycparser-20231013-5387-1wj1ewr/pycparser-2.21
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz @ file:///private/tmp/python-pytz-20231024-5128-1jupfew/pytz-2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.1
readme-renderer==42.0
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
rpds-py==0.10.6
scipy==1.11.3
selenium==4.16.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soccerdata==1.5.3
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.50
stack-data==0.6.3
streamlit==1.28.0
tenacity==8.2.3
termcolor==2.3.0
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
trio==0.23.2
trio-websocket==0.11.1
trove-classifiers==2023.11.29
twine==4.0.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
undetected-chromedriver==3.5.4
unicode==2.9
Unidecode==1.3.7
urllib3==2.0.7
validators==0.22.0
virtualenv==20.25.0
wcwidth==0.2.9
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
xattr==0.10.1
yarl==1.9.2
zipp==3.17.0
